Inline Case files for MaxLineLengthSpec
