--- detekt-rules-style/src/test/resources/MaxLineLengthWithLongComments.kt
@@ -1,31 +1,0 @@
-package cases
-
-@Suppress("unused")
-class MaxLineLengthWithLongComments {
-    // Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua.
-    /* Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua. */
-
-    /*
-     * Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua.
-     */
-    companion object {
-        val LOREM_IPSUM = "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua."
-    }
-
-    val loremIpsumField = "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua."
-
-    fun main() {
-        val thisIsAVeryLongValName = "This is a very, very long String that will break the MaxLineLength"
-
-        if (thisIsAVeryLongValName.length > "This is not quite as long of a String".length) {
-            println("It's indeed a very long String")
-        }
-
-        val loremIpsum = getLoremIpsum()
-
-        println(loremIpsum)
-
-    }
-
-    fun getLoremIpsum() = "Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam voluptua."
-}
